Experimental Session 4
I managed to manipulate the material of an object whenever it is hit by
a bullet. I created the health for the enemy, so when it reaches 0, the
player's score increases by a number. I set some additional gui
elements. I implemented the first enemy actions, such as simple
movement, and the ability to destroy the player when collides with him.