Experimental Session 5
I used a switch for the player to be able to change between the bullet
and a new weapon called missle. I made the missle to apply more damage
to the enemy when it hits, however it reduces the ammo amount at a
quicker rate. I started experimenting with the enemy ai target detection
and some basic movements.