Experimental Session 8
I created a function for the enemy to shoot at a certain time interval
when it starts following the player. The enemy will be able to kill the
player, either by shooting him or crashing into him. I changed the mode
of the camera from first person to third person where an angulated
camera follows the player from the top in third person. I also managed
to make the head and the arm follow the mouse to shoot, whilst movement
and rotation can be achieved with the keyboard.